Restructure code for better cohesion.
This commit introduces the class `LocalDesignDocument` to contain both the file
handling and the json handling and validation of the local design documents.
The existing class `LocalDesignDocumentsSelector` no longer returns instances
of `java.io.File`, but rather, returns instances of this new class. The other
changes in this commit are just making every thing work with the new signature.